Update and rename sma_trader_app.py to app.py
Rather big update: Renamed to a more pythonic name. Separated the trading function from price and kline/candles to make it easier to control errors with the added advantage of not having to restart script on errors, it should now reconnect the socket with error and keep going, which preserves all the trading quantities and dividends. Changed how dividends is calculated and stored so all pairs have access to the same amount of base currency. Added a 'cancel order if it hasn't sold/bought before opposite signal comes function'. Made reading last trade more reliable by saving to .CSV, so that even if the app crashes it can refer to last trade. Moved profit to the top constant to set for all pairs instead of by having it individually. Test mode with buy/sell option. Added a 'number_of_days' option to set how much history to download for easier config in the history function.